make don id uint declaration consistent

+++ contracts/src/v0.8/keystone/test/mocks/CapabilityConfigurationContract.sol
@@ -7,7 +7,7 @@ import {ERC165} from "../../../vendor/openzeppelin-solidity/v4.8.3/contracts/uti
 contract CapabilityConfigurationContract is ICapabilityConfiguration, ERC165 {
   mapping(uint256 => bytes) private s_donConfiguration;
 
-  function getCapabilityConfiguration(uint256 donId) external view returns (bytes memory configuration) {
+  function getCapabilityConfiguration(uint32 donId) external view returns (bytes memory configuration) {
     return s_donConfiguration[donId];
   }
